+++ posthog/src/main/java/com/posthog/internal/PostHogRemoteConfig.kt
@@ -1,6 +1,7 @@
 package com.posthog.internal
 
 import com.posthog.PostHogConfig
+import com.posthog.PostHogInternal
 import com.posthog.PostHogOnFeatureFlags
 import com.posthog.internal.PostHogPreferences.Companion.FEATURE_FLAGS
 import com.posthog.internal.PostHogPreferences.Companion.FEATURE_FLAGS_PAYLOAD
@@ -18,7 +19,8 @@ import java.util.concurrent.atomic.AtomicBoolean
  * @property api the API
  * @property executor the Executor
  */
-internal class PostHogRemoteConfig(
+@PostHogInternal
+public class PostHogRemoteConfig(
     private val config: PostHogConfig,
     private val api: PostHogApi,
     private val executor: ExecutorService,
@@ -53,7 +55,7 @@ internal class PostHogRemoteConfig(
      * Use this to notify listeners that cached surveys may have changed.
      */
     @Volatile
-    var onRemoteConfigLoaded: (() -> Unit)? = null
+    public var onRemoteConfigLoaded: (() -> Unit)? = null
 
     init {
         preloadSessionReplayFlag()
@@ -119,12 +121,12 @@ internal class PostHogRemoteConfig(
         }
     }
 
-    override fun loadRemoteConfig(
+    public fun loadRemoteConfig(
         distinctId: String,
         anonymousId: String?,
         groups: Map<String, String>?,
-        internalOnFeatureFlags: PostHogOnFeatureFlags?,
-        onFeatureFlags: PostHogOnFeatureFlags?,
+        internalOnFeatureFlags: PostHogOnFeatureFlags? = null,
+        onFeatureFlags: PostHogOnFeatureFlags? = null,
     ) {
         executor.executeSafely {
             if (config.networkStatus?.isConnected() == false) {
@@ -392,7 +394,7 @@ internal class PostHogRemoteConfig(
         }
     }
 
-    fun loadFeatureFlags(
+    public fun loadFeatureFlags(
         distinctId: String,
         anonymousId: String?,
         groups: Map<String, String>?,
@@ -560,6 +562,10 @@ internal class PostHogRemoteConfig(
     override fun getFeatureFlag(
         key: String,
         defaultValue: Any?,
+        distinctId: String?,
+        groups: Map<String, String>?,
+        personProperties: Map<String, String>?,
+        groupProperties: Map<String, String>?,
     ): Any? {
         // since we pass featureFlags as a value, we need to load it from cache if needed
         loadFeatureFlagsFromCacheIfNeeded()
@@ -570,6 +576,10 @@ internal class PostHogRemoteConfig(
     override fun getFeatureFlagPayload(
         key: String,
         defaultValue: Any?,
+        distinctId: String?,
+        groups: Map<String, String>?,
+        personProperties: Map<String, String>?,
+        groupProperties: Map<String, String>?,
     ): Any? {
         // since we pass featureFlags as a value, we need to load it from cache if needed
         loadFeatureFlagsFromCacheIfNeeded()
@@ -577,7 +587,12 @@ internal class PostHogRemoteConfig(
         return readFeatureFlag(key, defaultValue, featureFlagPayloads)
     }
 
-    override fun getFeatureFlags(): Map<String, Any>? {
+    override fun getFeatureFlags(
+        distinctId: String?,
+        groups: Map<String, String>?,
+        personProperties: Map<String, String>?,
+        groupProperties: Map<String, String>?,
+    ): Map<String, Any>? {
         val flags: Map<String, Any>?
         synchronized(featureFlagsLock) {
             flags = featureFlags?.toMap()
@@ -585,16 +600,16 @@ internal class PostHogRemoteConfig(
         return flags
     }
 
-    override fun isSessionReplayFlagActive(): Boolean = sessionReplayFlagActive
+    public fun isSessionReplayFlagActive(): Boolean = sessionReplayFlagActive
 
-    fun getRequestId(): String? {
+    public fun getRequestId(): String? {
         loadFeatureFlagsFromCacheIfNeeded()
         synchronized(featureFlagsLock) {
             return requestId
         }
     }
 
-    fun getFlagDetails(key: String): FeatureFlag? {
+    public fun getFlagDetails(key: String): FeatureFlag? {
         loadFeatureFlagsFromCacheIfNeeded()
 
         synchronized(featureFlagsLock) {
@@ -602,7 +617,7 @@ internal class PostHogRemoteConfig(
         }
     }
 
-    fun getSurveys(): List<Survey>? {
+    public fun getSurveys(): List<Survey>? {
         synchronized(remoteConfigLock) {
             return surveys
         }