+++ posthog/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- feat: `PostHogConfig` now accepts queue and remote config constructors ([#287])
+
 ## 3.22.0 - 2025-09-25
 
 - feat: Add a server-side stateless interface([#284](https://github.com/PostHog/posthog-android/pull/284))